Sometimes when a page loads the weather icon doesn't show because there's no URL. A closer inspection into this reveals that the weather API occassionally produces the wrong results. For example, sometimes it says that I'm in China, even though I'm in the US. When that happens there's no icons.
Also, when the page loads it takes at least a few seconds for the weather API to return the user the relevant data. In the meantime the user will see an incomplete UI. To prevent this I created a loading page, and its CSS attributes, that will appear until API values (celsius and fahrenheit) appear in the Main component props.

Will need to move it to its own file as its next step!